+++ src/main/java/com/emulator/jmme/agbemu/hardware/GameBoyMemory.java
@@ -6,21 +6,21 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class GameBoyMemory implements Memory {
-    private Map<Integer, Short> addresses;
+    private Map<Short, Byte> addresses;
 
     public GameBoyMemory() {
         this.addresses = new HashMap<>();
 
         for (int i = 0; i <= 0xFFFF; i++) {
-            addresses.put(i, (short)0);
+            addresses.put((short)i, (byte)0);
         }
     }
 
-    public Map<Integer, Short> getAddresses() {
+    public Map<Short, Byte> getAddresses() {
         return addresses;
     }
 
-    public void setAddresses(Map<Integer, Short> addresses) {
+    public void setAddresses(Map<Short, Byte> addresses) {
         this.addresses = addresses;
     }
 }
